去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@22797 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsMgr.cpp
@@ -1071,7 +1071,7 @@ namespace engine
       // the msg is not mine, dispatch to sub object
       // restore the type
       INT32 type = (INT32) msg->TID ;
-      UINT32 opCode = (UINT32)(msg->opCode) ;
+      INT32 opCode = msg->opCode ;
       msg->TID = 0 ;
 
       if ( CLS_REPL == type ||
